+++ src/main/webapp/ui/src/Inventory/components/DetailedListing/Card.js
@@ -28,7 +28,7 @@ import SearchContext from "../../../stores/contexts/Search";
 import useStores from "../../../stores/use-stores";
 import { globalStyles, type Theme } from "../../../theme";
 import clsx from "clsx";
-import ImagePreview from "../../components/ImagePreview";
+import ImagePreview from "../../../components/ImagePreview";
 import { StyledMenu } from "../../../components/StyledMenu";
 import Portal from "@mui/material/Portal";
 import NavigateContext from "../../../stores/contexts/Navigate";
